feat: adapt for didc array numer replaced Uint8Array
