+++ packages/nns/src/utils/converter.utils.ts
@@ -3,31 +3,6 @@ import { Buffer } from "buffer";
 // @ts-ignore (no type definitions for crc are available)
 import crc from "crc";
 
-export const uint8ArrayToBigInt = (array: Uint8Array): bigint => {
-  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
-  if (typeof view.getBigUint64 === "function") {
-    return view.getBigUint64(0);
-  } else {
-    const high = BigInt(view.getUint32(0));
-    const low = BigInt(view.getUint32(4));
-
-    return (high << BigInt(32)) + low;
-  }
-};
-
-export const arrayBufferToUint8Array = (buffer: ArrayBuffer): Uint8Array =>
-  new Uint8Array(buffer);
-
-export const uint8ArrayToArrayOfNumber = (array: Uint8Array): Array<number> =>
-  Array.from(array);
-
-export const arrayOfNumberToUint8Array = (numbers: Array<number>): Uint8Array =>
-  new Uint8Array(numbers);
-
-export const asciiStringToByteArray = (text: string): Array<number> => {
-  return Array.from(text).map((c) => c.charCodeAt(0));
-};
-
 export const toHexString = (bytes: Uint8Array) =>
   bytes.reduce((str, byte) => str + byte.toString(16).padStart(2, "0"), "");
