+++ packages/utils/src/index.ts
@@ -1,5 +1,6 @@
 export type { CanisterOptions } from "./types/canister.options";
 export * from "./utils/actor.utils";
 export * from "./utils/agent.utils";
+export * from "./utils/array.utils";
 export * from "./utils/asserts.utils";
 export * from "./utils/did.utils";
